Added _safe_ns_register()
If we register a prefix => namespace pair with XML::LibXML that is
already in use at the document level, our pair is ignored
without a error message of any kind.

This code checks to see if a namespace URI is already in use by the
document before setting the mapping.

This makes sure that the prefix is unique.

If we don't use this code, the XPath matches can succeed for nodes
with the wrong (document defined namespace).

See http://www.simplicidade.org/notes/archives/2008/08/xmllibxml_brain.html
for more info.

Signed-off-by: Pedro Melo <melo@simplicidade.org>