Added subscribe() and deliver_messages() APIs to Net::SAPO::Broker
The subscribe() method has the standard set of parameters, common to
Protocol::SAPO::Broker.

The deliver_messages() API accepts a single integer parameter, a timeout.

It will wait for new messages for `timeout` seconds. If a message is
received, it is delivered via the callback given on the subscribe() call.

The timeout is then reset and Net::SAPO::Protocol will wait for another
message.

The deliver_messages() API will return if a period of `timeout` seconds
is elapsed without receiving any messages.

Signed-off-by: Pedro Melo <melo@simplicidade.org>