Major rework of Protocol::SAPO::Broker to clean up the internal logic
The main goal was to implement auto reconnect.

So all the logic of connection establishement is moved to four method:

 * _connect;
 * _disconnect;
 * _lost_connection;
 * _reconnect.

The on_message callback no longer passes the Message element, only the
full XPath object. Also the Fault element was removed from the fault hook,
you have the full XPath object only.

All methods check the current state. They will die if the method is not
supported on that state. For example, calling disconnect() if state is
not 'connected', will die.

There where a lot of hooks that where called immediatly after
a state change. Given that a state change to X will always call the
`on_state_X` hook, removed those extra hooks:

 * connected: use on_state_connected;
 * connect_error: use on_state_connect_error;
 * eof: use on_state_eof;
 * write_error: use on_state_write_error;
 * read_error: use on_state_read_error.

The read_error and write_error hooks croak if the error is not cleared by
a on_state_{write,read}_error.

We also revamped a lot of the parameter checking code:

 * all callbacks on_* require a CODE ref;
 * on_send and on_connect are required to create a Protocol:: object.

Regardin the error API, we splited the _clear_error to a new method.
Previously you would call _set_error() with undef.

All return values from hooks are now explicitly ignored.

A DESTROY method was added: it calls disconnect if a connection is active.

Signed-off-by: Pedro Melo <melo@simplicidade.org>